Implement Reader and Writer using a single file
Add very basic implementation of Reader and Writer. They just read and
write file with fixed name in a given directory. In the future Writer
will calculate the checksum and write 2 files - data and sum. Reader
will read both and verify sum on Close.

This commit also create a filesystem abstraction which is needed for unit
testing, but also to decouple code from os package.