Add store option to disable integrity check
This can be really useful for development purposes. Developer might edit data file directly using his editor of choice and load the edited version into the program.

There is also a chance that production data need to be updated. In such
cases it is better to alter only a specific version of file instead of
disabiling integrity check globally. This can be done by replacing checksum file content with specical word "ALTERED"